Main page layout

+++ database/migrations/2020_07_12_194757_add_column_invite_id_to_users_table.php
@@ -14,7 +14,7 @@ class AddColumnInviteIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('invite_id')->nullable();
+            $table->unsignedBigInteger('invite_id')->nullable();
 
             $table->foreign('invite_id')
                 ->references('id')->on('invites')
